Concerns about blocking io in the right places
For vague reasons we defined Generator.generate to return
a future although all the implementations does is blocking
io in a very imperative fashion. We then went on to juggle
futures like crazy clowns without much reason.

I have now removed those, and made sure the blocking call
to generate is wrapped in a future and executed on a separate
context for blocking io in the controller. Presto - much cleaner
code.